flask_backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
numpy==1.24.3
pandas==2.1.4
pillow==10.1.0
tensorflow==2.16.1
keras>=3.3,<4
python-dotenv==1.0.1
supabase==2.6.0
httpx==0.27.2
pybars3==0.9.7
